Auto-Dungeon: Load waypoint onPlayerPositionChange() and toggleDungeonName.isOn()
Auto-Dungeon:
Auto-Dungeon now loads ONLY when toggleDungeonName.isOn() AND tile == Dungeon tile.

This fixes the issue where, after a dungeon is complete, or switching floors, the CaveBot toggles off, causing your character to stop botting.

This also makes it so normal waypoint loading doesn't interact with dungeon waypoint loading. So, if botting normal waypoints and the floor is changed, the waypoint or recorder will not stop.

Libraries:
Added a few new libraries to populate as time goes on. Items, NPCs, Strings, Tables, and Tiles. More libraries as well as information about the libraries coming soon.

Targeting:
With the introduction of the "Strings" library, targeting now is able to prioritize targets based on creature level. If the creature level == nil, the priority does not change.

With that being said, a separate targeting priority list will be implemented if a boss monster is introduced.